logger moved to external package.
