+++ cmd/safetyculture-exporter/cmd/utils/validation.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"github.com/SafetyCulture/safetyculture-exporter/pkg/logger"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
@@ -10,7 +11,7 @@ var lgr *zap.SugaredLogger
 // Check the error and exit the process when not nil
 func Check(err error, msg string) {
 	if lgr == nil {
-		lgr = GetLogger()
+		lgr = logger.GetLogger()
 	}
 
 	if err != nil {
