+++ pkg/logger/logger_test.go
@@ -1,15 +1,15 @@
-package util_test
+package logger_test
 
 import (
+	"github.com/SafetyCulture/safetyculture-exporter/pkg/logger"
 	"testing"
 
-	util "github.com/SafetyCulture/safetyculture-exporter/cmd/safetyculture-exporter/cmd/utils"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestGetLogger_should_return_same_instance_every_time(t *testing.T) {
-	t1 := util.GetLogger()
-	t2 := util.GetLogger()
+	t1 := logger.GetLogger()
+	t2 := logger.GetLogger()
 
 	assert.Equal(t, t1, t2)
 }
